+++ src/sentry/static/sentry/app/views/groupSimilar/groupSimilarView.jsx
