+++ src/sentry/static/sentry/app/views/groupSimilar/mergedItem.jsx
