+++ src/sentry/static/sentry/app/views/groupSimilar/mergedList.jsx
