+++ src/sentry/static/sentry/app/views/groupSimilar/mergedToolbar.jsx
