+++ src/sentry/static/sentry/app/views/groupSimilar/similarItem.jsx
