+++ src/sentry/static/sentry/app/views/groupSimilar/similarList.jsx
