+++ src/sentry/static/sentry/app/views/groupSimilar/similarToolbar.jsx
